requirements.txt
python-telegram-bot==20.7
python-dotenv
python-binance
gspread
oauth2client
notion-client
pandas
apscheduler
cryptography
aiohttp